+++ pkgs/flutter_http_example/lib/main.dart
@@ -15,6 +15,14 @@ import 'package.dart';
 
 void main() {
   runApp(Provider<Client>(
+      // `Provider` calls its `create` argument once when a `Client` is
+      // first requested (through `BuildContext.read<Client>()`) and uses that
+      // same instance for all future requests.
+      //
+      // Reusing the same `Client` may:
+      // - reduce memory usage
+      // - allow caching of fetched URLs
+      // - allow connections to be persisted
       create: (_) => http_factory.httpClient(),
       child: const PackageSearchApp(),
       dispose: (_, client) => client.close()));