+++ pkgs/cronet_http/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.9.1-wip
+
+* Update example to fetch and display scores of package:cronet_http from pub.dev.
+
 ## 1.9.0
 
 * Add `CronetEngine.startNetLogToFile` and `CronetEngine.stopNetLog`.